+++ Changelog.md
@@ -10,6 +10,7 @@
 - Changed the cart fixed and percent discount promotion action types to only act on carts, not on cart items
 - Added the `priority` field to the `products` table
 - Added the `priority` field to the `master_products` table
+- Added default sorting to `ProductSearch`: by priority, descending
 - Added the static `getAvailableWithoutCoupon()` and `getAvailableOnes()` methods to the `Promotion` model (but not to the interface)
 - Added the `active()` and `notDepeleted()` scopes to the `Promotion` model
 - Added the `PromotionEvent` interface